apps/hotel-api/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
PyMySQL
PyJWT
requests
sqlalchemy
python-multipart
bcrypt